requirements.txt
streamlit==0.82.0
pandas==1.0.5
numpy==1.18.5
Pillow==8.3.0
scikit_learn==0.24.2
